requirements.txt
arabic-reshaper==2.0.15
astroid==2.3.3
astunparse==1.6.3
autopep8==1.4.4
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
colorclass==2.2.0
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
esprima==4.0.1
et-xmlfile==1.0.1
freetype-py==2.1.0.post1
future==0.18.2
gevent==1.4.0
gitdb2==2.0.6
GitPython==3.0.5
glfw==1.8.7
greenlet==0.4.15
idna==2.8
imageio==2.6.1
imageio-ffmpeg==0.3.0
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
jdcal==1.4.1
jedi==0.15.2
json-tricks==3.13.5
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
matplotlib==3.1.2
mccabe==0.6.1
moviepy==1.0.1
msgpack==0.6.2
msgpack-numpy==0.4.4.3
numexpr==2.7.0
numpy==1.18.0
opencv-python==4.1.2.30
openpyxl==3.0.2
pandas==0.25.3
parso==0.5.2
pexpect==4.7.0
pickleshare==0.7.5
Pillow==7.0.0
pip-check==2.6
proglog==0.1.9
prompt-toolkit==3.0.2
psutil==5.6.7
PsychoPy==3.2.4
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.19
pyglet==1.4.9
Pygments==2.5.2
pylint==2.4.4
PyOpenGL==3.1.5
pyOpenSSL==19.1.0
pyosf==1.0.5
pyparallel==0.2.2
pyparsing==2.4.6
PyQt5==5.14.0
PyQt5-sip==12.7.0
pyserial==3.4
python-bidi==0.4.2
python-dateutil==2.8.1
python-gitlab==1.15.0
pytz==2019.3
PyYAML==5.2
pyzmq==18.1.1
questplus==2019.4
requests==2.22.0
scipy==1.4.1
six==1.13.0
smmap2==2.0.5
sounddevice==0.3.14
SoundFile==0.10.3.post1
tables==3.6.1
terminaltables==3.1.0
tqdm==4.41.1
traitlets==4.3.3
typed-ast==1.4.0
urllib3==1.25.7
wcwidth==0.1.8
wrapt==1.11.2
wxPython==4.0.7.post2
xarray==0.14.1
xlrd==1.2.0
